# Cache milagro g1 pubkeys