--- a/parachain/rustfmt.toml
+++ b/parachain/rustfmt.toml
@@ -21,3 +21,4 @@ match_block_trailing_comma = true
 trailing_comma = "Vertical"
 trailing_semicolon = false
 use_field_init_shorthand = true
+edition = "2021"