for SNO-356 and PRs accompanied
https://github.com/Snowfork/incubator-milagro-crypto-rust/pull/2
https://github.com/Snowfork/milagro_bls/pull/2
https://github.com/Snowfork/cumulus/pull/19