@yrong 2023-04-21
# Benchmark

- execution time decreased by almost **4 times less** than before as following.

|Extrinsic  | Minimum Execution Time before optimization(us) | Minimum Execution Time after optimization(us)  |  
|-----------------------|-----------------------|-------------------------|
|sync_committee_period_update | 125_394                                |     138_632 
|import_finalized_header| 119_688                                |  30_960 
|import_execution_header | 119_457                                  | 30_946

- storage proof increased by almost **3 times more** than before as following.

|Extrinsic  | Storage Proof Size Estimated before optimization(bytes) | Storage Proof Size Estimated after optimization(bytes)  |  
|-----------------------|-----------------------|-------------------------|
|sync_committee_period_update | 72217                                |     207289  
|import_finalized_header| 43615                                |  111151 
|import_execution_header | 37729                                  | 105265