@vgeddes 2023-04-25
> @vgeddes In recent [commit](https://github.com/Snowfork/snowbridge/pull/811/commits/15ddc13b4882ffeac76330ea451c6795cbed1f28) I've added the  optimization about `substract absent G1 points` as we discussed before,  benchmark [here](https://github.com/Snowfork/cumulus/pull/19/commits/4a75a5a562ac3062008f6a1ad0c4dd104866a370) shows only a minor improvement achieved(kind of unexpected). Seems not a key point and I'll double check.

Ah, too bad. Let's keep the optimization. It does shave off some
Weight. 👍