@yrong 2023-05-04
@doubledup For this PR to work we need to upgrade rust toolchain to `nightly-2023-04-22`,  I would suspect Nix setup will just respect config [here](https://github.com/Snowfork/snowbridge/blob/912de31bd6cbff60fa4550226b43f5f9f2b61e82/parachain/rust-toolchain.toml#L9) and no special handling required, but correct me if I'm wrong.

@vgeddes The only issue left to introduce `generic-const-array` is we need to add `--release` for unit test or it will just failed,  CI changed [here](https://github.com/Snowfork/snowbridge/blob/912de31bd6cbff60fa4550226b43f5f9f2b61e82/.github/workflows/parachain.yml#L102) accordingly.  I would assume still related to rust compiler and should be fine for us.