@yrong 2023-05-04
@vgeddes The only issue left is we need to add `--release` for unit test or it will just failed,  CI changed [here](https://github.com/Snowfork/snowbridge/blob/912de31bd6cbff60fa4550226b43f5f9f2b61e82/.github/workflows/parachain.yml#L102) accordingly.  I would assume still related to rust compiler and not a block issue.