@vgeddes 2023-05-04
> @doubledup For this PR to work we need to upgrade rust toolchain to `nightly-2023-04-22`, I would suspect Nix setup will just respect config [here](https://github.com/Snowfork/snowbridge/blob/912de31bd6cbff60fa4550226b43f5f9f2b61e82/parachain/rust-toolchain.toml#L9) and no special handling required, but correct me if I'm wrong.

Nope, I tried this, and we first need to update the `rust-overlay` input package for our Nix env:
```bash
nix flake lock --update-input rust-overlay
```